Cleanup [React] Replace createRef across the code
