+++ src/App.js
@@ -59,7 +59,7 @@ import 'igz-controls/scss/common.scss'
 import './scss/main.scss'
 import { createPortal } from 'react-dom'
 
-import Notification from './common/Notification/Notification'
+import Notifications from './common/Notifications/Notifications'
 
 const Page = lazyRetry(() => import('./layout/Page/Page'))
 const CreateJobPage = lazyRetry(() => import('./components/CreateJobPage/CreateJobPage'))
@@ -303,7 +303,7 @@ const App = () => {
         <Suspense fallback={<LoaderForSuspenseFallback />}>
           <RouterProvider router={router} />
         </Suspense>
-        {createPortal(<Notification />, document.getElementById('overlay_container'))}
+        {createPortal(<Notifications />, document.getElementById('overlay_container'))}
       </div>
     </div>
   )
