--- src/common/Breadcrumbs/breadcrums.scss
@@ -1,49 +1,0 @@
-@import '~igz-controls/scss/mixins';
-@import '~igz-controls/scss/colors';
-
-.breadcrumbs {
-  @include resetSpaces;
-
-  font-size: 20px;
-  line-height: 23px;
-
-  &__list {
-    @include resetSpaces;
-
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    font-weight: 400;
-    letter-spacing: 0.00938em;
-    list-style-type: none;
-  }
-
-  &__item {
-    position: relative;
-    max-width: 700px;
-    color: $topaz;
-
-    &:last-child {
-      color: $mulledWine;
-      font-weight: 500;
-      font-size: 20px;
-      line-height: 23px;
-    }
-  }
-
-  &__separator {
-    display: flex;
-    margin: 0 8px;
-    color: $topaz;
-    transition: transform 0.2s linear;
-    user-select: none;
-
-    &_tumbler {
-      cursor: pointer;
-    }
-
-    &_active {
-      transform: rotate(90deg);
-    }
-  }
-}
