+++ src/common/Notifications/notifications.scss
@@ -1,0 +1,8 @@
+.notifications-wrapper {
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  display: flex;
+  flex: 1 1;
+  flex-direction: column;
+}
