+++ src/constants.js
@@ -642,6 +642,9 @@ export const CHART_TYPE_LINE = 'line'
 export const CHART_TYPE_BAR = 'bar'
 
 /*=========== ARTIFACTS LIMITS =============*/
-export const ARTIFACT_MAX_CHUNK_SIZE = 1048576       // 1MB
-export const ARTIFACT_MAX_PREVIEW_SIZE = 10485760    // 10MB
-export const ARTIFACT_MAX_DOWNLOAD_SIZE = 104857600  // 100MB
+export const ARTIFACT_MAX_CHUNK_SIZE = 1048576 // 1MB
+export const ARTIFACT_MAX_PREVIEW_SIZE = 10485760 // 10MB
+export const ARTIFACT_MAX_DOWNLOAD_SIZE = 104857600 // 100MB
+
+/*=========== NOTIFICATION =============*/
+export const NOTIFICATION_DURATION = 500
